+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Added
 - Guild commands for testing
 
+### Changed
+- `./src/modules/wordle.js` new game function to work with first new command
+
 ## [Alpha 1.2.1](https://github.com/Nevah5/wordlebot/releases/tag/a1.2.1) - 06.02.2022
 ### Added
 - Help embed